buffer size INT_MAX

+++ Circle1/get_next_line/srcs/get_next_line_bonus.c
@@ -6,7 +6,7 @@
 /*   By: jikoo <jikoo@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/06 19:45:23 by jikoo             #+#    #+#             */
-/*   Updated: 2022/09/13 16:33:16 by jikoo            ###   ########.fr       */
+/*   Updated: 2022/09/13 18:17:17 by jikoo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -79,7 +79,7 @@ char	*get_next_line(int fd)
 		}
 		return (NULL);
 	}
-	buffer = (char *)malloc(sizeof(char) * (BUFFER_SIZE + 1));
+	buffer = (char *)malloc(sizeof(char) * ((size_t)BUFFER_SIZE + 1));
 	if (!buffer)
 		return (NULL);
 	if (!(backup[fd]))
